+++ pkg/resources/base/base.go
@@ -28,21 +28,10 @@ import (
 
 const (
 	componentName                 = "common"
-	istiodName                    = "istiod"
-	istiodPilotName               = "istiod-pilot"
 	istioReaderName               = "istio-reader"
 	istioReaderServiceAccountName = "istio-reader-service-account"
-	istiodServiceAccountName      = "istiod-service-account"
 	IstioConfigMapName            = "istio"
 )
-
-var istiodLabel = map[string]string{
-	"app": istiodName,
-}
-
-var pilotLabel = map[string]string{
-	"app": "pilot",
-}
 
 var istioReaderLabel = map[string]string{
 	"app": istioReaderName,
@@ -70,13 +59,8 @@ func (r *Reconciler) Reconcile(log logr.Logger) error {
 
 	for _, res := range []resources.Resource{
 		r.serviceAccountReader,
-		r.serviceAccount,
-		r.role,
-		r.clusterRole,
 		r.clusterRoleReader,
 		r.clusterRoleBindingReader,
-		r.clusterRoleBinding,
-		r.roleBinding,
 		r.configMap,
 	} {
 		o := res()
